feat: Moved Redis to its own project
