+++ platform/Platform.sln
@@ -82,6 +82,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Platform.Search", "src\abst
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Platform.Infrastructure", "src\abstractions\Platform.Infrastructure\Platform.Infrastructure.csproj", "{25B60DB3-6ABF-4F76-B60D-73381A88FD09}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Platform.Cache", "src\abstractions\Platform.Cache\Platform.Cache.csproj", "{F763848C-9E85-4FED-85C2-790C5E58E56A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -144,6 +146,10 @@ Global
 		{25B60DB3-6ABF-4F76-B60D-73381A88FD09}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25B60DB3-6ABF-4F76-B60D-73381A88FD09}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25B60DB3-6ABF-4F76-B60D-73381A88FD09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F763848C-9E85-4FED-85C2-790C5E58E56A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F763848C-9E85-4FED-85C2-790C5E58E56A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F763848C-9E85-4FED-85C2-790C5E58E56A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F763848C-9E85-4FED-85C2-790C5E58E56A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -165,5 +171,6 @@ Global
 		{23B45FED-62AA-4C53-9DF5-67511CD66B8E} = {D52455F1-3BF5-46F4-9833-633817248295}
 		{F5525998-DB69-4209-8832-E3176C0DBC6D} = {D52455F1-3BF5-46F4-9833-633817248295}
 		{25B60DB3-6ABF-4F76-B60D-73381A88FD09} = {D52455F1-3BF5-46F4-9833-633817248295}
+		{F763848C-9E85-4FED-85C2-790C5E58E56A} = {D52455F1-3BF5-46F4-9833-633817248295}
 	EndGlobalSection
 EndGlobal
