+++ platform/src/apis/Platform.Api.Insight/Platform.Api.Insight.csproj
@@ -17,9 +17,8 @@
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
             <CopyToPublishDirectory>Never</CopyToPublishDirectory>
         </None>
-        <PackageReference Include="AspNetCore.HealthChecks.Redis"/>
-        <PackageReference Include="Microsoft.Azure.StackExchangeRedis"/>
         <PackageReference Include="Microsoft.Azure.WebJobs"/>
+        <ProjectReference Include="..\..\abstractions\Platform.Cache\Platform.Cache.csproj"/>
         <ProjectReference Include="..\..\abstractions\Platform.Functions\Platform.Functions.csproj"/>
         <PackageReference Include="AspNetCore.HealthChecks.SqlServer"/>
         <PackageReference Include="Dapper.SqlBuilder"/>
